app/requirements.txt
appdirs==1.4.4
APScheduler==3.6.3
astroid==2.4.2
attrs==20.3.0
autobahn==20.12.3
Automat==20.2.0
beautifulsoup4==4.9.3
boltons==20.2.1
botocore==1.19.59
bs4==0.0.1
certifi==2020.12.5
cffi==1.14.4
chardet==4.0.0
click==7.1.2
constantly==15.1.0
cryptography==3.3.1
cssselect==1.1.0
dateparser==1.0.0
decorator==4.4.2
eventkit==0.8.6
face==20.1.1
fake-useragent==0.1.11
fastapi==0.63.0
forex-python==1.5
h11==0.12.0
hyperlink==21.0.0
ib-insync==0.9.64
idna==2.10
incremental==17.5.0
isort==5.7.0
jmespath==0.10.0
lazy-object-proxy==1.4.3
lxml==4.6.2
mccabe==0.6.1
nest-asyncio==1.4.3
newtulipy==0.4.4
numpy==1.19.5
pandas==1.2.1
pandas-datareader==0.9.0
parse==1.19.0
pipupgrade==1.8.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
pydantic==1.7.3
pyee==8.1.0
PyHamcrest==2.0.2
pylint==2.6.0
pyOpenSSL==20.0.1
pyppeteer==0.2.5
pyquery==1.4.3
python-binance==0.7.9
python-dateutil==2.8.1
python-decouple==3.4
python-telegram-bot==13.1
pytz==2020.5
regex==2020.11.13
requests==2.25.1
requests-html==0.10.0
rope==0.18.0
s3transfer==0.3.4
service-identity==18.1.0
simplejson==3.17.2
six==1.15.0
soupsieve==2.1
starlette==0.13.6
toml==0.10.2
tornado==6.1
tqdm==4.56.0
Twisted==20.3.0
txaio==20.12.1
tzlocal==2.1
ujson==4.0.1
urllib3==1.26.2
uvicorn==0.13.3
w3lib==1.22.0
websockets==8.1
wrapt==1.12.1
yahoo-fin==0.8.6
zope.interface==5.2.0
